requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
flask-wtf
flask-login
flask-sqlalchemy
flask-bcrypt
email_validator
gunicorn
python-decouple